Make DoubleBuffer work with empty batches
DoubleBuffer.empty() method was used to see if there are more batches
to process. However, since it would work solely by looking into the
buffer's length, that meant that DoubleBuffer could not detect that
there's an empty batch to process. This lead to the dead lock when the
node was waiting acknowledgement from the client that the batch was
received, but the client didn't know that there was a batch, never
acknowledging this.

Fixes #62