Always resume RecordStream fiber when user issues stop
Previously, if `RecordStream` was suspended, issuing a `stop`
from a controller would resume it only if is waiting for the node
to deliver more data. This however doesn't work if user suspended
the request, and while suspended, they try to stop it (one example
of this situation would be where the user stops the request when
receiving node_disconnected, which by definition gets dispatched
asynchronously from the request).

Now the RecordStream is resumed if it's waiting for the records
from the node or if it's suspended explicitly.