Replace `Const!` with `const`
D1 support was dropped in v15.x.x. This commit replaces the use of
`Const!` with D2 builtin `const` because `Const!` was only used to
transition to D2, i.e., it enabled us to D2 features while supporting
both D1 and D2. All usages of `Const!` in this code base are replaced.
Note, there are no uses of `Immut!` or `Inout!` in the code base.

Fixes sociomantic-tsunami/dlsproto#113.